requirements.txt
Django==2.2.10
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
